pyproject.toml
[project]
name = "dropbox-version-history-movie"
version = "0.1.0"
description = "Generate a video showing the changes to an image file using your Dropbox history"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "dropbox>=12.0.2",
    "moviepy>=2.1.1",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
]
